Stepped closer to solution
Purple bugs during brush operation, were caused by memory error. Fact is, FIFO volume of 30 samples allowed the CODEC to finish its five-clock memory write cycle. Moreover, it stopped exactly on 4-th action: writing data package to RAM. Before the CODEC was going to do it, the fifo became empty, which made WRITE_ENABLE signal to stay high, and entire FSM - to disable. After a while, display turned on , and the address bus took a different value. In combination with enabled WE and the leftovers in write registers, it caused extra write operation to random memory cells.
This error and a pile of others, made me take a decision to simplify memory CODEC. Although, it means rebuilding it in 7th(or 8th) time,  I believe this is the last (in 6th or 7th time).